Fix issue #28
Created two packages (src, test), one for source code and one for unit testing.
Created an yaml file for input data.
Created two classes inside the src package (cmv, decide).
Added the code for reading the input data in the main file.
Moved the code from main file to cmv class in the src package.
Added the code for cmv3 method from the branch cmv-3.
Added return statements and argument checking for cmv3 method.
Renamed the names of the methods with cmv_methodNumber.
Organized and added the necessary arguments to the existing methods in the cmv class.
Added comments for bugs in cmv class.
Added the docs directory that will generate the documentation for our code.