Fix issue#43
Fixed a bug inside the cmv 9 method by adding a counter for the while loop and fixing the first if statement inside the while loop.
Added the documentation for the cmv 9 method.
Created four test cases (positive, negative, 2 tests with invalid data) for the cmv 9 method and added the documentation for them.